feat(protocol): add Adapter types namespace
