+++ packages/protocol/src/index.ts
@@ -6,6 +6,7 @@ export * from "./sink/index.js";
 export * from "./bus/index.js";
 export * from "./event/index.js";
 export * from "./notification/index.js";
+export * from "./adapter/index.js";
 export * from "./plan/index.js";
 export * from "./team/index.js";
 export * from "./ingress/index.js";
